Changed project structure
change project structure to add in UI components
