+++ setup.py
@@ -1,18 +1,7 @@
 from configparser import ConfigParser
 from cryptography.fernet import Fernet
+from app.ui.colors import BColors
 import sqlite3
-
-
-class BColors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
 
 print('Writing config...')
 config = ConfigParser()
